Work on updated provider
