+++ src/components/ProviderInitializer/helpers/index.ts
@@ -1,3 +1,2 @@
 export * from './getOperaProvider';
 export * from './getExtensionProvider';
-export * from './getHwWalletProvider';
